refactor: MeetingCommand 패키지 수정
