+++ src/main/java/com/mobble/mobbleserver/application/meeting/port/provided/MeetingCreatePort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.meeting.port.provided;
 
-import com.mobble.mobbleserver.application.meeting.command.request.CreateMeetingCommand;
+import com.mobble.mobbleserver.application.meeting.command.CreateMeetingCommand;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
 
 public interface MeetingCreatePort {
