+++ src/main/java/com/mobble/mobbleserver/application/meeting/port/provided/MeetingUpdatePort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.meeting.port.provided;
 
-import com.mobble.mobbleserver.application.meeting.command.request.UpdateMeetingCommand;
+import com.mobble.mobbleserver.application.meeting.command.UpdateMeetingCommand;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
 
 public interface MeetingUpdatePort {
